refactor: move files back and use 'imports' on 'py_binary'
