+++ gazelle/python/BUILD.bazel
@@ -48,11 +48,13 @@ go_library(
 py_binary(
     name = "helper",
     srcs = [
-        "//rules_python_gazelle_helper:__main__.py",
-        "//rules_python_gazelle_helper:parse.py",
-        "//rules_python_gazelle_helper:std_modules.py",
+        "__main__.py",
+        "parse.py",
+        "std_modules.py",
     ],
-    main = "//rules_python_gazelle_helper:__main__.py",
+    # This is to make sure that the current directory is added to PYTHONPATH
+    imports = ["."],
+    main = "__main__.py",
     visibility = ["//visibility:public"],
 )
