--- gazelle/rules_python_gazelle_helper/BUILD.bazel
@@ -1,6 +1,0 @@
-# gazelle:ignore
-
-exports_files(
-    srcs = glob(["*.py"]),
-    visibility = ["//python:__pkg__"],
-)
